further deconstruction of tables in docx to allow granular control of formatting
