+++ src/paradoc/__init__.py
@@ -1,3 +1,4 @@
-from .document import MY_DOCX_TMPL, MY_DOCX_TMPL_BLANK, OneDoc
+from .common import MY_DOCX_TMPL, MY_DOCX_TMPL_BLANK
+from .document import OneDoc
 
 __all__ = [MY_DOCX_TMPL, MY_DOCX_TMPL_BLANK, OneDoc]
