+++ src/paradoc/document.py
@@ -4,26 +4,14 @@ import os
 import pandas as pd
 import pathlib
 import shutil
-from dataclasses import dataclass
 
 import pypandoc
-from docx import Document
-from docxcompose.composer import Composer
 from typing import Dict
 from .concepts import Table
 from .formatting import TableFormat
-from .utils import close_word_docs_by_name, docx_update, get_list_of_files
+from .utils import get_list_of_files
 
-MY_DOCX_TMPL = pathlib.Path(__file__).resolve().absolute().parent / "resources" / "template.docx"
-MY_DOCX_TMPL_BLANK = pathlib.Path(__file__).resolve().absolute().parent / "resources" / "template_blank.docx"
-
-
-@dataclass
-class MdFile:
-    path: pathlib.Path
-    is_appendix: bool
-    new_file: pathlib.Path
-    build_file: pathlib.Path
+from .common import ExportFormats, MarkDownFile
 
 
 class OneDoc:
@@ -53,7 +41,7 @@ class OneDoc:
     def __init__(
         self,
         source_dir=None,
-        export_format="docx",
+        export_format=ExportFormats.DOCX,
         main_prefix="00-main",
         app_prefix="01-app",
         clean_build_dir=True,
@@ -95,7 +83,7 @@ class OneDoc:
             new_file = self.build_dir / md_file.relative_to(self.source_dir).with_suffix(".docx")
             build_file = self.build_dir / md_file.relative_to(self.source_dir)
 
-            md_file = MdFile(
+            md_file = MarkDownFile(
                 path=md_file,
                 is_appendix=is_appendix,
                 new_file=new_file,
@@ -114,11 +102,6 @@ class OneDoc:
             shutil.rmtree(self.build_dir, ignore_errors=True)
 
     def compile(self, output_name, auto_open=False, metadata_file=None):
-        from .formatting import Formatting
-        from .formatting.utils import (
-            apply_custom_styles_to_docx,
-            fix_headers_after_compose
-        )
         from .utils import variable_sub
 
         dest_file = (self.dist_dir / output_name).with_suffix(f".{self.export_format}").resolve().absolute()
@@ -163,84 +146,17 @@ class OneDoc:
                 filters=["pandoc-crossref"],
                 encoding="utf8",
             )
+        if self.export_format == ExportFormats.DOCX:
+            from paradoc.io.word.exporter import WordExporter
 
-        # Main Document - Format Style
-        composer_main = Composer(Document(MY_DOCX_TMPL))
-        composer_main.doc.add_page_break()
-
-        for i, md in enumerate(self.md_files_main):
-            doc_in = Document(str(md.new_file))
-            doc_in.add_page_break()
-            composer_main.append(doc_in)
-            logging.info(f"Added {md.new_file}")
-
-        main_format = Formatting(False, self.paragraph_style_map, self.table_format)
-        _ = self._reformat_doc(composer_main.doc, False)
-        composer_main.doc.add_page_break()
-
-        # Appendix - Format Style
-        composer_app = Composer(Document(MY_DOCX_TMPL_BLANK))
-        for i, md in enumerate(self.md_files_app):
-            doc_in = Document(str(md.new_file))
-            doc_in.add_page_break()
-            composer_app.append(doc_in)
-            logging.info(f"Added {md.new_file}")
-
-        app_paragraph_style = dict()
-        app_paragraph_style.update(self.appendix_heading_map)
-        app_paragraph_style.update(self.paragraph_style_map)
-
-        app_format = Formatting(True, app_paragraph_style, self.table_format)
-        _ = self._reformat_doc(composer_main.doc, True)
-
-        composer_main.append(composer_app.doc)
-
-        fix_headers_after_compose(composer_main.doc)
-
-        print("Close Existing Word documents")
-        close_word_docs_by_name([output_name, f"{output_name}.docx"])
-
-        print(f'Saving Composed Document to "{dest_file}"')
-        composer_main.save(dest_file)
-
-        docx_update(str(dest_file))
+            wordx = WordExporter(self)
+            wordx.convert_to_docx(output_name, dest_file)
 
         if auto_open is True:
             os.startfile(dest_file)
 
     def add_table(self, name, df: pd.DataFrame, caption: str, tbl_format: TableFormat = None):
         self.tables[name] = Table(name, df, caption, tbl_format)
-
-    def _reformat_doc(self, doc: Document, is_appendix, style_doc=None):
-        from paradoc import MY_DOCX_TMPL
-        from paradoc.utils import iter_block_items
-        from docx.table import Table as DocxTable
-        from docx.text.paragraph import Paragraph
-        from .formatting.utils import format_table, format_paragraph, format_captions, get_table_ref
-
-        document = style_doc if style_doc is not None else Document(MY_DOCX_TMPL)
-        prev_table = False
-        refs = dict()
-
-        for block in iter_block_items(doc):
-            if type(block) == Paragraph:
-                if prev_table and len(block.runs) > 0:
-                    block.runs[0].text = "\n" + block.runs[0].text
-                    prev_table = False
-                    block.paragraph_format.space_before = None
-                if block.style.name in ("Image Caption", "Table Caption"):
-                    ref_ = format_captions(block, is_appendix)
-                    refs.update(ref_)
-                else:
-                    format_paragraph(block, document, self.paragraph_style_map)
-
-            elif type(block) == DocxTable:
-                tbl_source = get_table_ref(block, self.tables)
-                if tbl_source is not None and tbl_source.format is not None:
-                    format_table(block, document, tbl_source.format)
-                prev_table = True
-
-        return refs
 
     @property
     def main_dir(self):
